Extract the validation of TypeMapImpl to a class
To simplify the TypeMapImpl implemenation, and for single responsibility
principle, extract the functionality of validate and get unmapped
properties to another class. And also provides an extensible configurable
validator in the future if need it.
